// src/types/product.ts
export interface Product {
    id: number;
    name: string;
    slug: string;
    price: number;
    price_formatted: string;
    image: string;
    diskon: string;
    weight: number;
    type: number;
    preorder: boolean;
    mulai: string | null;
    selesai: string | null;
    sisastok: number;
    sisastok_label: number;
    jenispotongan: string | null;
    potongan: number | null;
    image_url: string;
    final_price: number;
    final_price_formatted: string;
    video_available: boolean;
    video: string | null;
}